Добавляет компоненты в проект
