+++ src/components/card-color-bar.js
@@ -1,0 +1,7 @@
+export const getCardColorBarTemplate = () => (
+  `<div class="card__color-bar">
+    <svg class="card__color-bar-wave" width="100%" height="10">
+      <use xlink:href="#wave"></use>
+    </svg>
+  </div>`
+);
