+++ src/components/load-more-button.js
@@ -1,0 +1,1 @@
+export const getLoadButtonTemplate = () => `<button class="load-more" type="button">load more</button>`;
